Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int radius;
+    float radius;
     const pi=3.14;
-    scanf("%d",&radius);
-    printf("Area: %d",pi*r*r);
+    scanf("%f",&radius);
+    printf("Area: %f",pi*r*r);
     return 0;
 }
